Refactor as utils package
