+++ pyecharts/utils/__init__.py
@@ -3,20 +3,16 @@ from __future__ import unicode_literals
 
 import codecs
 import os
-import operator
-import sys
-
-PY2 = sys.version_info[0] == 2
 
 
 def get_resource_dir(*paths):
     """
-
-    :param path:
+    Return absolute path for a directory or file inside the project.
+    :param paths: paths relative to project root dir.
     :return:
     """
-    current_path = os.path.dirname(__file__)
-    resource_path = os.path.join(current_path, *paths)
+    project_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+    resource_path = os.path.join(project_dir, *paths)
     return resource_path
 
 
@@ -79,45 +75,3 @@ def merge_js_dependencies(*chart_or_name_list):
     # items which should be included in front part.
     fol = [x for x in front_optional_items if x in fist_items]
     return front_must_items + fol + dependencies
-
-
-# ------- Lazy Module ------#
-
-def proxy_method(func):
-    def inner(self, *args):
-        if self._wrapped is None:
-            self._setup()
-        return func(self._wrapped, *args)
-
-    return inner
-
-
-class LazyObject(object):
-    _wrapped = None
-
-    def __init__(self, func):
-        self.__dict__['_setupfunc'] = func
-
-    __getattr__ = proxy_method(getattr)
-
-    def __setattr__(self, key, value):
-        if key == '_wrapped':
-            self.__dict__['_wrapped'] = value
-        else:
-            if self._wrapped is None:
-                self._setup()
-            setattr(self._wrapped, key, value)
-
-    def _setup(self):
-        self._wrapped = self._setupfunc()
-
-    __getitem__ = proxy_method(operator.getitem)
-
-    if PY2:
-        __str__ = proxy_method(str)
-        __unicode__ = proxy_method(unicode)  # NOQA: unicode undefined on PY3
-        __nonzero__ = proxy_method(bool)
-    else:
-        __bytes__ = proxy_method(bytes)
-        __str__ = proxy_method(str)
-        __bool__ = proxy_method(bool)
